requirements.txt
decorator==4.4.1
Django==3.0.5
djangorestframework==3.12.1
djongo==1.3.3
dnspython==1.16.0
pymongo==3.10.1
regex==2020.6.8
requests==2.22.0
